Move Brush class into its own file
